% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ana.R
\name{gene_sturcture}
\alias{gene_sturcture}
\title{Obtain the chromosomal and genetic information of the species from the protein file}
\usage{
gene_sturcture(protein_df, taxonomyID)
}
\arguments{
\item{protein_df}{a data frame from function \code{\link{get_protein}} output}

\item{taxonomyID}{taxonomy ID of your species}
}
\value{
a list include two data frame(One contains chromosomal information
    and the other contains genetic information )
}
\description{
Obtain the chromosomal and genetic information of the species from the protein file
}
\author{
Yuan Zhen
}
