% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{get_pfm}
\alias{get_pfm}
\title{Get the pfm ID from the gene family name you gave}
\usage{
get_pfm(protein_familyname)
}
\arguments{
\item{protein_familyname}{The name of the gene family}
}
\value{
pfm IDs
}
\description{
Get the pfm ID from the gene family name you gave
}
\seealso{
\code{\link{get_taxonomy}} which can be used to get taxonomy ID

\code{\link{get_protein}} which can be used to get taxonomy ID
}
\author{
Zhen Yuan
}
