% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_ana.R
\name{get_protein}
\alias{get_protein}
\title{Protein families are obtained from NCBI by PFM number and taxonomy ID}
\usage{
get_protein(pfm = "PF02701", taxonomyID = "3983")
}
\arguments{
\item{pfm}{pfm ID}

\item{taxonomyID}{taxonomy ID}
}
\value{
A data frame with protein accession ids ,descriptions,
sourcedb,taxonomy ID and organism
}
\description{
Protein families are obtained from NCBI by PFM number and taxonomy ID
}
\examples{

\dontrun{

my_pfm=get_pfm(protein_familyname="Dof")

taxonomy_id=get_taxonomy(organism_name="soybean")

###Please note that you may have multiple outputs for these two results, so choose the one that suits you best.

protein_info=get_protein(pfm=my_pfm,taxonomyID=taxonomy_id)

}


}
\seealso{
\code{\link{get_pfm}} which can be used to get pfm ID

\code{\link{get_taxonomy}} which can be used to get taxonomy ID
}
\author{
Zhen Yuan
}
