% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{get_taxonomy}
\alias{get_taxonomy}
\title{Get the taxonomy ID from the species name you gave}
\usage{
get_taxonomy(organism_name)
}
\arguments{
\item{organism_name}{}
}
\value{
taxonomy IDs
}
\description{
Get the taxonomy ID from the species name you gave
}
\seealso{
\code{\link{get_protein}} which can be used to get taxonomy ID

\code{\link{get_pfm}} which can be used to get pfm ID
}
\author{
Zhen Yuan
}
