% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ana.R
\name{plot_genes}
\alias{plot_genes}
\title{Draw the position of genes on chromosomes}
\usage{
plot_genes(gene_plot)
}
\arguments{
\item{gene_plot}{a list from function \code{\link{gene_sturcture}} output}
}
\description{
Draw the position of genes on chromosomes
}
