% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro_ana.R
\name{protein_properties}
\alias{protein_properties}
\title{Information on the physicochemical properties of proteins is obtained}
\usage{
protein_properties(pro_seq)
}
\arguments{
\item{pro_seq}{a AAStringset from function \code{\link{protein_seq}} output}
}
\description{
Information on the physicochemical properties of proteins is obtained
}
