% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro_ana.R
\name{protein_seq}
\alias{protein_seq}
\title{Obtain sequence information for proteins}
\usage{
protein_seq(protein_df)
}
\arguments{
\item{protein_df}{a data frame from function \code{\link{get_protein}} output}
}
\description{
Obtain sequence information for proteins
}
