% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro_ana.R
\name{str2_sting}
\alias{str2_sting}
\title{Turn NCBI obtained information into DNAString or AAString objects}
\usage{
str2_sting(seq, type = "dna")
}
\arguments{
\item{seq}{sequence information obtained from NCBI}

\item{type}{Select the type you need to deal with.("dna" or "pro")}
}
\description{
Turn NCBI obtained information into DNAString or AAString objects
}
